+++ Algolia.Search/Models/Responses/AddObjectResponse.cs
@@ -30,7 +30,7 @@ namespace Algolia.Search.Models.Responses
     public class AddObjectResponse
     {
         public DateTime CreatedAt { get; set; }
-        public int TaskID { get; set; }
+        public long TaskID { get; set; }
         public string ObjectID { get; set; }
     }
 }
